Move UsernamePickerTestCase to test_login
It's not particularly OIDC-specific, so I think it belongs here.
